src/TextGen_refactored/requirements.txt
# Core dependencies
jinja2>=3.1.2
pyyaml>=6.0
typing-extensions>=4.5.0

# Language model providers
openai>=1.3.0
anthropic>=0.5.0
google-generativeai>=0.3.0
# No need for separate Groq library - we use requests directly

# UI framework
streamlit>=1.27.0
pandas>=2.0.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.28.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0 